requirements.txt
black
flask
python_dotenv
openai>1